✨ add route for sca/events
We want to use a proxy route to do our request as we are storing the
api key (https://grafana.com/developers/plugin-tools/create-a-plugin/extend-a-plugin/add-authentication-for-data-source-plugins#authenticate-using-the-data-source-proxy)

We do not want the API key to be stored on the frontend so instead it's
stored as a secret on the graphana server side. This saves us from
writing a custom backend but still keeping the secrets out of the
frontend.

There is only one path at the moment.